// ui/fields/number/value/normalize-number-value/contract/types.ts
/** NumberValueOptions описывает данные публичного контракта своего владельца. */
export type NumberValueOptions = Readonly<{
  min?: number | undefined
  max?: number | undefined
  softMin?: number | undefined
  softMax?: number | undefined
  step?: number | undefined
}>